@@ -14,7 +14,9 @@
     "lint-fix": "eslint . --fix --ext js,jsx,ts,tsx",
     "postinstall": "husky install",
     "storybook": "start-storybook -s ./public -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "prettier": "npx prettier src __tests__ --check",
+    "prettier:fix": "npm run prettier -- --write"
   },
   "repository": {
     "type": "git",
@@ -24,15 +26,19 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.1",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^13.3.0",
-    "axios": "^0.27.2",
-    "husky": "^8.0.1",
+    "@types/jest": "^29.2.5",
+    "axios": "^1.2.2",
+    "husky": "8.0.3",
     "moment": "^2.29.4",
-    "next": "12",
+    "next": "^13.1.1",
+    "next-redux-wrapper": "^8.1.0",
     "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
     "react-toastify": "^9.0.8",
     "sass": "^1.45.2",
     "sass-loader": "^13.0.2",
@@ -54,29 +60,32 @@
     "@types/react": "^18.0.17",
     "@types/react-beautiful-dnd": "^13.1.2",
     "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.8.1",
-    "@typescript-eslint/parser": "^5.8.1",
-    "babel-jest": "^28.1.3",
-    "babel-loader": "^8.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "babel-jest": "^29.3.1",
+    "babel-loader": "^9.1.2",
     "css-loader": "^6.5.1",
     "eslint": "^8.6.0",
-    "eslint-config-next": "^12.0.7",
+    "eslint-config-next": "^13.1.1",
+    "eslint-plugin-react": "^7.32.2",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^28.1.3",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
     "jest-mock-axios": "^4.5.0",
-    "lint-staged": "^13.0.3",
+    "lint-staged": "^13.1.0",
     "local-ssl-proxy": "^1.3.0",
-    "msw": "^0.44.2",
+    "msw": "^0.49.2",
     "npm-run-all": "^4.1.5",
     "pinst": "^3.0.0",
     "react-hooks-testing-library": "^0.6.0",
     "react-test-renderer": "^18.2.0",
     "style-loader": "^3.3.1",
-    "ts-jest": "^28.0.8",
-    "webpack": "^5.65.0"
+    "ts-jest": "^29.0.3",
+    "webpack": "^5.76.0",
+    "whatwg-fetch": "^3.6.2"
   },
   "engines": {
-    "node": "16.x"
+    "node": "18.x"
   },
   "lint-staged": {
     "*.+(ts|tsx)": [
@@ -85,5 +94,9 @@
   },
   "msw": {
     "workerDirectory": "public"
+  },
+  "volta": {
+    "node": "18.14.2",
+    "yarn": "1.22.19"
   }
 }
