@@ -1,2 +1,2 @@
 declare module '*.scss';
-import "@testing-library/jest-dom/extend-expect";
+import '@testing-library/jest-dom/extend-expect';
