@@ -1,36 +1,36 @@
 module.exports = {
-  collectCoverageFrom: [
-    "**/*.{js,jsx,ts,tsx}",
-    "!**/*.d.ts",
-    "!**/node_modules/**",
-  ],
-  moduleNameMapper: {
-    /* Handle CSS imports (with CSS modules)
+	collectCoverageFrom: [
+		'**/*.{js,jsx,ts,tsx}',
+		'!**/*.d.ts',
+		'!**/node_modules/**',
+	],
+	moduleNameMapper: {
+		/* Handle CSS imports (with CSS modules)
     https://jestjs.io/docs/webpack#mocking-css-modules */
-    "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy",
+		'^.+\\.module\\.(css|sass|scss)$': 'identity-obj-proxy',
 
-    // Handle CSS imports (without CSS modules)
-    "^.+\\.(css|sass|scss)$": "<rootDir>/__mocks__/styleMock.js",
+		// Handle CSS imports (without CSS modules)
+		'^.+\\.(css|sass|scss)$': '<rootDir>/__mocks__/styleMock.js',
 
-    /* Handle image imports
+		/* Handle image imports
     https://jestjs.io/docs/webpack#handling-static-assets */
-    "^.+\\.(jpg|jpeg|png|gif|webp|svg)$": "<rootDir>/__mocks__/fileMock.js",
+		'^.+\\.(jpg|jpeg|png|gif|webp|svg)$': '<rootDir>/__mocks__/fileMock.js',
 
-    // Handle absolute imports
-    "^@/(.*)$": "<rootDir>/src/$1",
-  },
-  testPathIgnorePatterns: ["<rootDir>/node_modules/", "<rootDir>/.next/"],
-  testEnvironment: "jsdom",
-  transform: {
-    /* Use babel-jest to transpile tests with the next/babel preset
+		// Handle absolute imports
+		'^@/(.*)$': '<rootDir>/src/$1',
+	},
+	testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/.next/'],
+	testEnvironment: 'jsdom',
+	transform: {
+		/* Use babel-jest to transpile tests with the next/babel preset
     https://jestjs.io/docs/configuration#transform-objectstring-pathtotransformer--pathtotransformer-object */
-    "^.+\\.(js|jsx|ts|tsx)$": ["babel-jest", { presets: ["next/babel"] }],
-    "^.+\\.tsx?$": "ts-jest",
-  },
-  transformIgnorePatterns: [
-    "/node_modules/",
-    "^.+\\.module\\.(css|sass|scss)$",
-  ],
-  setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
-  testRegex: "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
+		'^.+\\.(js|jsx|ts|tsx)$': ['babel-jest', { presets: ['next/babel'] }],
+		'^.+\\.tsx?$': 'ts-jest',
+	},
+	transformIgnorePatterns: [
+		'/node_modules/',
+		'^.+\\.module\\.(css|sass|scss)$',
+	],
+	setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+	testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
 };
