@@ -1,2 +1,2 @@
 export const CHALLENGES_SCREEN = 'Challenges';
-export const DUMMY_PROFILE = 'dummyProfile.png';
+export const DUMMY_PROFILE = '/dummyProfile.png';
