@@ -3,17 +3,17 @@ import fetch from '@/helperFunctions/fetch';
 import { toast, ToastTypes } from '@/helperFunctions/toast';
 
 const userData = async () => {
-  try {
-    const { requestPromise } = fetch({
-      url: USER_SELF,
-      method: 'get',
-    });
-    const response = await requestPromise;
-    return response.data;
-  } catch (error) {
-    toast(ToastTypes.ERROR, (error as Error).message);
-    return error;
-  }
+    try {
+        const { requestPromise } = fetch({
+            url: USER_SELF,
+            method: 'get',
+        });
+        const response = await requestPromise;
+        return response.data;
+    } catch (error) {
+        toast(ToastTypes.ERROR, (error as Error).message);
+        return error;
+    }
 };
 
 export default userData;
