@@ -1,40 +1,52 @@
 import { useState, useEffect } from 'react';
 import fetch from '@/helperFunctions/fetch';
 
-const useFetch = (url: string, options: object = {}) => {
-  const [response, setResponse] = useState<any>({});
-  const [error, setError] = useState<any>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  useEffect(() => {
-    let cancel: () => void;
-    (async () => {
-      setIsLoading(true);
-      try {
-        const { requestPromise, cancelApi } = fetch({
-          url,
-          method: 'get',
-          ...options,
-        });
-        cancel = cancelApi;
-        const fetchPromise = await requestPromise;
-        setResponse(fetchPromise.data);
-      } catch (err) {
-        setError(err);
-      } finally {
-        setIsLoading(false);
-      }
-    })();
+type EmptyFunction = () => void;
 
-    return () => {
-      cancel();
-      setError(null);
-      setResponse({});
-      setIsLoading(true);
+const useFetch = (url: string, options: object = {}, isCall = true) => {
+    const [response, setResponse] = useState<any>({});
+    const [error, setError] = useState<any>(null);
+    const [isLoading, setIsLoading] = useState(true);
+    const callAPI = async (): Promise<any | EmptyFunction> => {
+        let cancel: () => void;
+        setIsLoading(true);
+        try {
+            const { requestPromise, cancelApi } = fetch({
+                url,
+                method: 'get',
+                ...options,
+            });
+            cancel = cancelApi;
+            const fetchPromise = await requestPromise;
+            setResponse(fetchPromise.data);
+            return cancel;
+        } catch (err) {
+            setError(err);
+        } finally {
+            setIsLoading(false);
+        }
     };
-  }, [url]);
-  return {
-    response, error, isLoading,
-  };
+    useEffect(() => {
+        let cancel: any;
+        if (isCall) {
+            cancel = callAPI();
+        }
+
+        return () => {
+            if (typeof cancel == 'function') {
+                cancel();
+            }
+            setError(null);
+            setResponse({});
+            setIsLoading(true);
+        };
+    }, [url, isCall]);
+    return {
+        response,
+        error,
+        isLoading,
+        callAPI,
+    };
 };
 
 export default useFetch;
