@@ -1,20 +1,20 @@
 import task from '@/interfaces/task.type';
 
 export type dragDropProps = {
-  idleMembers: Array<string>,
-  unAssignedTasks: Array<task>,
-  refreshData:any,
+    idleMembers: Array<string>;
+    unAssignedTasks: Array<task>;
+    refreshData: any;
 };
 
 export type droppableComponent = {
-  droppableId: string,
-  idleMembers: Array<string>,
-  unAssignedTasks: Array<task>,
-  isTaskOnDrag:boolean,
+    droppableId: string;
+    idleMembers: Array<string>;
+    unAssignedTasks: Array<task>;
+    isTaskOnDrag: boolean;
 };
 
 export type draggableProps = {
-  draggableId: string,
-  index: number,
-  title?: string,
+    draggableId: string;
+    index: number;
+    title?: string;
 };
