@@ -1,6 +1,6 @@
 type award = {
-  dinero: number,
-  neelam: number,
+    dinero: number;
+    neelam: number;
 };
 
 export default award;
