@@ -1,23 +1,38 @@
 import award from '@/interfaces/award.type';
 
 type task = {
-  id: string,
-  title: string,
-  purpose: string,
-  featureUrl: string,
-  type: string,
-  links: string[],
-  endsOn: string,
-  startedOn: string,
-  status: string,
-  assignee?: string,
-  percentCompleted: number,
-  dependsOn: string[],
-  participants?: string[],
-  completionAward: award,
-  lossRate: award,
-  isNoteworthy: boolean,
-  createdBy: string,
+    id: string;
+    title: string;
+    purpose: string;
+    featureUrl: string;
+    type: string;
+    links: string[];
+    endsOn: string;
+    startedOn: string;
+    status: string;
+    assignee?: string;
+    percentCompleted: number;
+    dependsOn: string[];
+    participants?: string[];
+    completionAward: award;
+    lossRate: award;
+    isNoteworthy: boolean;
+    createdBy: string;
 };
+
+enum Tab {
+    ASSIGNED = 'ASSIGNED',
+    COMPLETED = 'COMPLETED',
+    AVAILABLE = 'AVAILABLE',
+    IN_PROGRESS = 'IN_PROGRESS',
+    NEEDS_REVIEW = 'NEEDS_REVIEW',
+    IN_REVIEW = 'IN_REVIEW',
+    VERIFIED = 'VERIFIED',
+    MERGED = 'MERGED',
+}
+
+const TABS = Object.values(Tab);
+
+export { TABS, Tab };
 
 export default task;
