@@ -1,20 +1,29 @@
 import { FC } from 'react';
 import type { AppProps } from 'next/app';
 import { ToastContainer } from 'react-toastify';
+import AppWrapperContext from '@/context';
+import IsUserAuthorizedContext from '@/context/isUserAuthorized';
+import { Provider } from 'react-redux';
+import { store } from '@/app/store';
 import 'react-toastify/dist/ReactToastify.css';
 import '@/styles/index.scss';
-import AppWrapperContext from '@/context'
 
 if (process.env.NEXT_PUBLIC_API_MOCKING === 'ON') {
-  require('../../__mocks__/mswInit');
+    require('../../__mocks__/index');
 }
 
-const MyApp: FC<AppProps> = ({ Component, pageProps }) => (
-  <AppWrapperContext>
-    <ToastContainer />
-    <Component {...pageProps} />
-  </AppWrapperContext>
-);
-
+const MyApp: FC<AppProps> = ({ Component, pageProps }) => {
+    const rdxStore = store();
+    return (
+        <Provider store={rdxStore}>
+            <AppWrapperContext>
+                <IsUserAuthorizedContext>
+                    <ToastContainer />
+                    <Component {...pageProps} />
+                </IsUserAuthorizedContext>
+            </AppWrapperContext>
+        </Provider>
+    );
+};
 
 export default MyApp;
